pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">= 3.12"
version = "0.2.1"
name = "hive_autocheck"
dependencies = [
    "argparse",
    "requests",
    "dataclasses",
    "pytest",
    "patool",
    "python-gitlab==2.10.0",
    "pydantic>=2.11.0",
    "pydantic-settings>=2.8.1",
    "wrapt>=1.17.0",
    "deprecated>=1.2.18",
]
readme = "README.md"
description = "A framework to easily create autocheck tests for Hive exercises"
classifiers = [
    "Private :: Do Not Upload",
]

[project.entry-points.pytest11]
hive_autocheck = "autocheck.fixtures"

[dependency-groups]
dev = [
    "mypy>=1.11.2,<2",
    "ruff>=0.11.2",
    "types-decorator>=5.2.0.20250324",
    "types-deprecated>=1.2.15.20250304",
    "types-requests",
]

[tool.ruff.lint]
select = [
    "ALL"
]

ignore = [
    "D",
    "S",       # flake8-bandit (security)
    "EM102",   # f-string-in-exception
    "G004",    # logging-f-string
    "TRY003",  # raise-vanilla-args
    "EM101",   # raw-string-in-exception
]

[tool.mypy]
plugins = ['pydantic.mypy']

# Ensure full coverage
disallow_untyped_calls = 1
disallow_untyped_defs = 1
disallow_incomplete_defs = 1
disallow_untyped_decorators = 1
check_untyped_defs = 1

# Restrict dynamic typing
disallow_any_generics = 1
disallow_subclassing_any = 1
warn_return_any = 1

# Know exactly what you're doing
warn_redundant_casts = 1
warn_unused_ignores = 1
warn_unused_configs = 1
warn_unreachable = 1
show_error_codes = 1

# Explicit is better than implicit
no_implicit_optional = 1

# Prevent modules from being skipped
ignore_missing_imports = 0
disallow_any_unimported = 1

